Rename tools to avoid conflicts
