+++ deploy/hubserving/ocr_rec/module.py
@@ -26,9 +26,9 @@ from paddlehub.module.module import moduleinfo, runnable, serving
 import cv2
 import paddlehub as hub
 
-from tools.infer.utility import base64_to_cv2
-from tools.infer.predict_rec import TextRecognizer
-from tools.infer.utility import parse_args
+from paddle_tools.infer.utility import base64_to_cv2
+from paddle_tools.infer.predict_rec import TextRecognizer
+from paddle_tools.infer.utility import parse_args
 from deploy.hubserving.ocr_rec.params import read_params
