+++ deploy/slim/prune/export_prune_model.py
@@ -31,7 +31,7 @@ from ppocr.modeling.architectures import build_model
 from ppocr.postprocess import build_post_process
 from ppocr.metrics import build_metric
 from ppocr.utils.save_load import load_model
-import tools.program as program
+import paddle_tools.program as program
 
 
 def main(config, device, logger, vdl_writer):
