+++ paddle_tools/export_center.py
@@ -29,7 +29,7 @@ from ppocr.modeling.architectures import build_model
 from ppocr.postprocess import build_post_process
 from ppocr.utils.save_load import load_model
 from ppocr.utils.utility import print_dict
-import tools.program as program
+import paddle_tools.program as program
 
 
 def main():
