+++ paddle_tools/export_model.py
@@ -28,7 +28,7 @@ from ppocr.modeling.architectures import build_model
 from ppocr.postprocess import build_post_process
 from ppocr.utils.save_load import load_model
 from ppocr.utils.logging import get_logger
-from tools.program import load_config, merge_config, ArgsParser
+from paddle_tools.program import load_config, merge_config, ArgsParser
 
 
 def export_single_model(model,
