+++ paddle_tools/infer/predict_det.py
@@ -25,7 +25,7 @@ import numpy as np
 import time
 import sys
 
-import tools.infer.utility as utility
+import paddle_tools.infer.utility as utility
 from ppocr.utils.logging import get_logger
 from ppocr.utils.utility import get_image_file_list, check_and_read
 from ppocr.data import create_operators, transform
