+++ paddle_tools/infer_kie_token_ser_re.py
@@ -38,8 +38,8 @@ from ppocr.utils.save_load import load_model
 from ppocr.utils.visual import draw_re_results
 from ppocr.utils.logging import get_logger
 from ppocr.utils.utility import get_image_file_list, load_vqa_bio_label_maps, print_dict
-from tools.program import ArgsParser, load_config, merge_config
-from tools.infer_kie_token_ser import SerPredictor
+from paddle_tools.program import ArgsParser, load_config, merge_config
+from paddle_tools.infer_kie_token_ser import SerPredictor
 
 
 class ReArgsParser(ArgsParser):
