+++ ppocr/data/imaug/copy_paste.py
@@ -20,7 +20,7 @@ from shapely.geometry import Polygon
 
 from ppocr.data.imaug.iaa_augment import IaaAugment
 from ppocr.data.imaug.random_crop_data import is_poly_outside_rect
-from tools.infer.utility import get_rotate_crop_image
+from paddle_tools.infer.utility import get_rotate_crop_image
 
 
 class CopyPaste(object):
